+++ umh-core/pkg/fsmv2/workers/communicator/transport/http/http_suite_test.go
@@ -1,0 +1,13 @@
+package transport_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestHTTPTransport(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "HTTP Transport Suite")
+}